implements the new low-level interface.
The old interface is preserved for now in as echelon.precursor and should be dissolve over time.
In addition, the new file format is introduced, which be dub v1.0.

This commit also resolves some long existing minor bugs, and adds a few minor features, like the access to the shuffle filter.